setup.py: Add env variable USE_CYTHON for compilation from pyx sources
pyx source files are now used only when USE_CYTHON is set to 1 or the
C source files have not yet been pre-compiled.

Only efl.eo.c is checked for this so if you cancel an initial build make
sure to enable USE_CYTHON when re-starting the build.

The pyx sources will now be included in the source tarball.