dbus_mainloop: reworked how we manage the dbus watch requests.
Dbus can request (via watch_add) more than one watch on the same fd (one
for write and one for read), but ecore cannot setup more than one fd_handler
on the same fd. So we need to manage this (weired) case internally, reusing
an existing fd_handler for more than one dbus watch.

This, at the end of the story, fix the ability to connect to the (fucking) PulseAudio
dbus server, that do not use the Session or System buses, but provide its own bus.