Optimize the _set_properties_from_keyword_args() function.
We do not need to check if the attribute exist to raise an exception, setattr() yet raise a well formatted expection in case of failure.

This way we avoid a dir() call and an IN check in a hot path.

The only drawback is that this change the raised exception in the case an attr not exist, from  AssertionError to AttributeError... I hope none was using that eception explicitly.