Use new cython cpdef-enums in evas
cpdef enums are automatically exported to C and
to python, no more need for the enums file and
the duplicated definition for py.

This is a cython 0.21 feature so from now on you will
need at least that version.

Require some <int> cast here and there, but a totally
win in the end.