Image.memfile_set: do not raise exceptions.
The first raise (when img do not support the buffer interface) is useless,
as python yet raise the same TypeError exception.

The latter is removed to be consistent with the docs and the C api,
so return a bool and do not rise the exception. We are going to move
to autogeneration sooner or later, so we must keep the api as much
as possible close to the C one, to minimize the future effort of porting
applications.