Evas.Smart: Comment out a warning for an issue we can't fix
When a smart event has event_info we need to convert it to a Python object.
We rely on the C library documentation to inform us when an event includes event_info
and assign a conversion function to the event handler based on this and pass the Python
object as an extra parameter to the callback. However when the documentation has omitted
this information we erroneously leave the extra parameter out, making the callbacks have
one less parameter, which our users rely on. Finding out the error after release is
useless since fixing it would mean breaking user callbacks.

This warning informed us at runtime when event_info was non-null but it's pretty much
useless so let's remove it.

Our smart callbacks should never have had two different signatures.